Adds WIP tests
