+++ package.json
@@ -19,6 +19,7 @@
     "mocha": "^2.3.3",
     "morgan": "~1.6.1",
     "serve-favicon": "~2.3.0",
-    "sqlite3": "^3.1.0"
+    "sqlite3": "^3.1.0",
+    "supertest": "^1.1.0"
   }
 }
